requirements.txt
pyyaml
munch
python-dotenv
openai
redis
jupyter_client
ipykernel
pandas
openpyxl
requests
bs4
googlesearch-python
inputimeout
matplotlib
datasets
cnocr
PyPDF2
pytube
pytubefix
python-docx
pillow
Bio
openai-whisper
anthropic
moviepy
python-pptx
pytesseract
loguru
mcp